requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-filter==22.1
django-restql==0.15.3
djangorestframework==3.14.0
djangorestframework-queryfields==1.0.0
drf-flex-fields==1.0.0
idna==3.4
Pillow==9.3.0
psycopg2==2.9.5
pycodestyle==2.10.0
pyPEG2==2.15.2
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.12
